requirements.txt
Pillow==6.1.0
beautifulsoup4==4.8.0
Django==1.11.1
pymysql==0.9.3
